End code when error rates converges to zero
The error rate of each pattern is collected into an array; if there is
one. At the end of each epoch, the sum of the squares of all of the
error rates collected is checked. If it equals 0, the “epochs loop” is
ended, and the rest of the code continues.